[Refactor] Move StreamLoadContext into ComputeEnv load
