+++ be/src/runtime/batch_write/batch_write_mgr.h
@@ -24,7 +24,6 @@
 #include "common/util/bthreads/executor.h"
 #include "runtime/batch_write/isomorphic_batch_write.h"
 #include "runtime/batch_write/txn_state_cache.h"
-#include "runtime/stream_load/stream_load_context.h"
 
 namespace brpc {
 class Controller;
@@ -36,6 +35,7 @@ class ExecEnv;
 class PStreamLoadRequest;
 class PStreamLoadResponse;
 class StreamLoadContext;
+class TUniqueId;
 class PUpdateTransactionStateRequest;
 class PUpdateTransactionStateResponse;
 class MetricRegistry;
