+++ be/src/runtime/rejected_record_sync_daemon.cpp
@@ -29,13 +29,13 @@
 #include "common/logging.h"
 #include "common/system/backend_options.h"
 #include "common/system/master_info.h"
+#include "compute_env/load/stream_load_context.h"
 #include "gen_cpp/HeartbeatService_types.h"
 #include "http/http_common.h"
 #include "platform/store_path.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/byte_buffer.h"
 #include "runtime/exec_env.h"
-#include "runtime/stream_load/stream_load_context.h"
 
 namespace starrocks {
