+++ be/src/runtime/stream_load/transaction_mgr.cpp
@@ -38,6 +38,7 @@
 #include "common/util/misc.h"
 #include "common/util/thrift_client_cache.h"
 #include "compute_env/load/load_stream_mgr.h"
+#include "compute_env/load/stream_load_context.h"
 #include "compute_env/load_path/base_load_path_mgr.h"
 #include "gen_cpp/FrontendService.h"
 #include "gen_cpp/FrontendService_types.h"
@@ -56,7 +57,6 @@
 #include "runtime/message_body_sink.h"
 #include "runtime/plan_fragment_executor.h"
 #include "runtime/stream_load/stream_context_mgr.h"
-#include "runtime/stream_load/stream_load_context.h"
 #include "runtime/stream_load/stream_load_executor.h"
 #include "runtime/stream_load/stream_load_metrics.h"
