+++ be/test/runtime/batch_write/batch_write_mgr_test.cpp
@@ -23,13 +23,13 @@
 #include "common/config_merge_commit_fwd.h"
 #include "common/thread/threadpool.h"
 #include "common/util/bthreads/executor.h"
+#include "compute_env/load/stream_load_context.h"
 #include "compute_env/load/time_bounded_stream_load_pipe.h"
 #include "gen_cpp/FrontendService.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "http/http_common.h"
 #include "http/http_headers.h"
 #include "runtime/exec_env.h"
-#include "runtime/stream_load/stream_load_context.h"
 #include "runtime/stream_load/stream_load_context_handle.h"
 
 namespace starrocks {
