+++ be/test/runtime/batch_write/isomorphic_batch_write_test.cpp
@@ -22,11 +22,11 @@
 #include "common/config_merge_commit_fwd.h"
 #include "common/thread/threadpool.h"
 #include "common/util/bthreads/executor.h"
+#include "compute_env/load/stream_load_context.h"
 #include "compute_env/load/time_bounded_stream_load_pipe.h"
 #include "gen_cpp/FrontendService.h"
 #include "http/http_common.h"
 #include "runtime/exec_env.h"
-#include "runtime/stream_load/stream_load_context.h"
 
 namespace starrocks {
